+++ autogen/src/lib.rs
@@ -1,0 +1,8 @@
+mod types;
+mod document;
+mod writer;
+
+pub use writer::*;
+pub use document::*;
+pub use types::*;
+
